backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pytesseract
pdf2image
openai
pypdf
python-dotenv
docxtpl
